{"schema":"decodex/commit/1","summary":"Split quantitative product manifest flow","authority":"manual"}
